+++ src/clustermergerpooltester.h
@@ -13,8 +13,8 @@ class ClusterMergerPoolTester : public ClusterMergerTester
 {
 public:
     ClusterMergerPoolTester(double zvalue);
-    bool shouldMerge(const std::vector<std::array<int,4>>& freq_c1,
-                     const std::vector<std::array<int,4>>& freq_c2);
+    bool shouldMerge(const std::vector<std::array<uint64_t,4>>& freq_c1,
+                     const std::vector<std::array<uint64_t,4>>& freq_c2);
 protected:
     std::unique_ptr<HypothesisTester> _tester;
 };
